+++ src/main/java/net/semanticmetadata/lire/solr/TopDocsIterator.java
@@ -1,0 +1,31 @@
+package net.semanticmetadata.lire.solr;
+
+import org.apache.htrace.fasterxml.jackson.databind.util.ArrayIterator;
+import org.apache.lucene.search.ScoreDoc;
+import org.apache.lucene.search.TopDocs;
+
+import java.util.Iterator;
+
+/**
+ * Created by mlux on 19.12.2016.
+ */
+public class TopDocsIterator implements Iterator<Integer> {
+    TopDocs docs;
+    ArrayIterator<ScoreDoc> innerIterator;
+
+    public TopDocsIterator(TopDocs docs) {
+        this.docs = docs;
+        innerIterator = new ArrayIterator<>(docs.scoreDocs);
+    }
+
+    @Override
+    public boolean hasNext() {
+        return innerIterator.hasNext();
+    }
+
+    @Override
+    public Integer next() {
+        ScoreDoc d = innerIterator.next();
+        return d.doc;
+    }
+}